create token provider

+++ Api/src/UrlShortener.Core/Base62EncodingExtensions.cs
@@ -5,7 +5,7 @@ public static class Base62EncodingExtensions
     private const string Base62Characters = "0123456789" +
                                              "ABCDEFGHIJKLMNOPQRSTUVWXYZ"+ 
                                               "abcdefghijklmnopqrstuvwxyz";
-    public static string EncodeToBase62(this int number)
+    public static string EncodeToBase62(this long number)
     {
         if (number == 0)
         {
@@ -15,7 +15,7 @@ public static class Base62EncodingExtensions
         var result = new Stack<char>();
         while (number > 0)
         {
-            result.Push(Base62Characters[number % 62]);
+            result.Push(Base62Characters[(int)number % 62]);
             number /= 62;
         }
